requirements.txt
discord.py
aiohttp
websokets
chardet
youtube_dl